Moving UserRolesManager plugin to security related bundle
