+++ Plugins/MsCrmTools.SecurityRelated/MsCrmTools.SecurityRelated.csproj
@@ -43,11 +43,13 @@
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.Composition" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Data" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.ServiceModel" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml" />
+    <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="PrivDiscover\AppCode\CommonDelegates.cs" />
@@ -67,14 +69,48 @@
       <DependentUpon>PrivDiscover.cs</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UserRolesManager\AppCode\ListViewItemComparer.cs" />
+    <Compile Include="UserRolesManager\AppCode\QueryHelper.cs" />
+    <Compile Include="UserRolesManager\AppCode\RoleAction.cs" />
+    <Compile Include="UserRolesManager\AppCode\RoleManager.cs" />
+    <Compile Include="UserRolesManager\AppCode\SystemUserManager.cs" />
+    <Compile Include="UserRolesManager\AppCode\ViewItem.cs" />
+    <Compile Include="UserRolesManager\AppCode\ViewManager.cs" />
+    <Compile Include="UserRolesManager\MainControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="UserRolesManager\MainControl.Designer.cs">
+      <DependentUpon>MainControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="UserRolesManager\Plugin.cs" />
+    <Compile Include="UserRolesManager\UserControls\PrincipalSelector.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="UserRolesManager\UserControls\PrincipalSelector.Designer.cs">
+      <DependentUpon>PrincipalSelector.cs</DependentUpon>
+    </Compile>
+    <Compile Include="UserRolesManager\UserControls\RoleSelector.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="UserRolesManager\UserControls\RoleSelector.Designer.cs">
+      <DependentUpon>RoleSelector.cs</DependentUpon>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Folder Include="RoleUpdater\" />
-    <Folder Include="UserRolesManager\" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="PrivDiscover\PrivDiscover.resx">
       <DependentUpon>PrivDiscover.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="UserRolesManager\MainControl.resx">
+      <DependentUpon>MainControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="UserRolesManager\UserControls\PrincipalSelector.resx">
+      <DependentUpon>PrincipalSelector.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="UserRolesManager\UserControls\RoleSelector.resx">
+      <DependentUpon>RoleSelector.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
