+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/AppCode/ListViewItemComparer.cs
