+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/AppCode/QueryHelper.cs
