+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/AppCode/RoleAction.cs
