+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/AppCode/ViewItem.cs
