+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/AppCode/ViewManager.cs
