+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/MainControl.resx
