+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/Plugin.cs
