+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/UserControls/PrincipalSelector.cs
