+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/UserControls/PrincipalSelector.resx
