+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/UserControls/RoleSelector.cs
