+++ Plugins/MsCrmTools.SecurityRelated/UserRolesManager/UserControls/RoleSelector.resx
